scribe: add watchdog for idle connections
Setup a watchdog for the Scribe connection. We don't want it
to keep it open forever. For instance, SLB VIP may trigger a RST
if idle more than a few minutes.

In the same time, deprecate the Scribe factory with explicit parameters
and use the EventTrackerConfig instead. Clients will benefit from
sane defaults for free.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>