http: limit the number of simultaneous requests
The HttpSender is invoked as callback from the diskspool writer
library.

In certain situations, when e.g. there is a very large backlog
of quarantined files and one tries to flush them all, we would run
out of file descriptors, as the sender tries to send them all
to the collector at the same time.

Be a little bit smarter and have a fixed-size pool of senders.
The default is 50 (under normal conditions, we shouldn't have more than
a few).

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>